+++ crates/mzoon/src/command.rs
@@ -1,0 +1,7 @@
+mod build;
+mod new;
+mod start;
+
+pub use build::build;
+pub use new::new;
+pub use start::start;
