+++ crates/mzoon/src/command/new.rs
@@ -1,0 +1,5 @@
+use anyhow::Result;
+
+pub async fn new(_project_name: String, _here: bool) -> Result<()> {
+    Ok(println!("Command `new` has not been implemented yet"))
+}
